[Modify] fetchRandomContent함수를 fetchGetFestival파일로 이동
# 219
